refactor: split component dialog
