+++ setup.cfg
@@ -1,0 +1,14 @@
+[metadata]
+name = smithpy
+version = 1.0.0
+description = Smith chart GUI tool
+long_description = file: README.md
+long_description_content_type = text/markdown
+
+[options]
+packages = find:
+package_dir =
+    =src
+
+[options.packages.find]
+where = src
