+++ src/smithpy/__init__.py
@@ -1,0 +1,5 @@
+"""Smith chart GUI package."""
+
+__all__ = ["__version__"]
+
+__version__ = "1.0.0"
