+++ src/smithpy/__main__.py
@@ -1,0 +1,5 @@
+"""Command line entry point for running the Smith chart GUI."""
+from .app import main
+
+if __name__ == "__main__":
+    main()
